update module to check weights on load, also fix docs and reorganize tests
